requirements.txt
streamlit
yfinance
ta
pandas
numpy
plotly
groq
feedparser
twstock
lxml
requests
gspread
python-dotenv
oauth2client
fastapi
uvicorn
pydantic
